Move http and error handlers to utils
