+++ .angular-cli.json
@@ -45,8 +45,6 @@
         "src/constants.ts",
         "src/mocks.ts",
         "src/test-data.ts",
-        "src/services/auth-http-helpers.ts",
-        "src/services/raven-error-handler.ts",
         "**/test.ts",
         "src/utils/*"
       ]
